Attempt to fix Travis CI Mac builds by installing libgit2
Checking on Travis CI on OSX and R-devel fails because git2r fails to
install when it tries to install devtools. The configure error is
'The libgit2 library that is required to build git2r was not found.'
It recommends to install libgit2 via homebrew and try again.
So, I added a bit to the .travis.yml to instruct it to do so,
though I'm not optimistic that it will fix the problem given that the
checks run fine on the other R versions on OSX.